+++ src/modules/dashboards/saga/index.ts
@@ -1,5 +1,6 @@
 import { saveDashboard } from './saveDashboard';
 import { cloneDashboard } from './cloneDashboard';
 import { viewDashboard } from './viewDashboard';
+import { resetDashboardFilters } from './resetDashboardFilters';
 
-export { viewDashboard, saveDashboard, cloneDashboard };
+export { viewDashboard, saveDashboard, cloneDashboard, resetDashboardFilters };
